+++ src/main/java/io/gate/gateapi/models/MyFuturesTrade.java
@@ -18,7 +18,6 @@ import com.google.gson.annotations.SerializedName;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import java.io.IOException;
-import java.math.BigDecimal;
 
 /**
  * MyFuturesTrade
@@ -30,7 +29,7 @@ public class MyFuturesTrade {
 
     public static final String SERIALIZED_NAME_CREATE_TIME = "create_time";
     @SerializedName(SERIALIZED_NAME_CREATE_TIME)
-    private BigDecimal createTime;
+    private Double createTime;
 
     public static final String SERIALIZED_NAME_CONTRACT = "contract";
     @SerializedName(SERIALIZED_NAME_CONTRACT)
@@ -120,7 +119,7 @@ public class MyFuturesTrade {
         this.id = id;
     }
 
-    public MyFuturesTrade createTime(BigDecimal createTime) {
+    public MyFuturesTrade createTime(Double createTime) {
         
         this.createTime = createTime;
         return this;
@@ -131,12 +130,12 @@ public class MyFuturesTrade {
      * @return createTime
     **/
     @javax.annotation.Nullable
-    public BigDecimal getCreateTime() {
+    public Double getCreateTime() {
         return createTime;
     }
 
 
-    public void setCreateTime(BigDecimal createTime) {
+    public void setCreateTime(Double createTime) {
         this.createTime = createTime;
     }
